Solution to P4c P4d

+++ part4/controllers/blogs.js
@@ -1,19 +1,42 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
   
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const blog = new Blog({ ...body, likes: body.likes || 0 })
+  if (!request.user) {
+    return response.status(401).json({ error: 'token invalid'})
+  }
+  const user = await User.findById(request.user)
+
+  const blogToAdd = {
+    ...body,
+    likes: body.likes || 0,
+    user: user._id
+  }
+  const blog = new Blog(blogToAdd)
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  if (!request.user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== request.user.toString()) {
+    return response.status(401).json({ error: 'unauthorized deletion of not-your blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
